bugfix: Don't try to delete the parent directory, if no feed entries were downloaded at all

+++ backup.py
@@ -645,6 +645,9 @@ class SitesBackup:
 			# end: for prefix
 		# end: for parent_id
 
+		if not os.path.exists(src_dir_parent): # no feed entries downloaded at all
+			return
+
 		print "\nRemoving the site's empty feeds content directory: %s" % \
 			(src_dir_parent)
 		os.rmdir(src_dir_parent)
